refactor: move util
