+++ src/NewPicker/PickerInput/RangePicker.tsx
@@ -1,3 +1,4 @@
+import { fillIndex } from '../../utils/miscUtil';
 import { useEvent, useMergedState } from 'rc-util';
 import useLayoutEffect from 'rc-util/lib/hooks/useLayoutEffect';
 import omit from 'rc-util/lib/omit';
@@ -19,7 +20,6 @@ import type {
 } from '../../interface';
 import type { PickerPanelProps } from '../PickerPanel';
 import PickerTrigger from '../PickerTrigger';
-import { fillIndex } from '../util';
 import PickerContext from './context';
 import useCellRender from './hooks/useCellRender';
 import useFieldsInvalidate from './hooks/useFieldsInvalidate';
