+++ src/NewPicker/PickerInput/hooks/useRangeValue.ts
@@ -1,10 +1,10 @@
 import { useEvent, useMergedState } from 'rc-util';
 import * as React from 'react';
 import type { GenerateConfig } from '../../../generate';
-import { formatValue, isSame, isSameTimestamp } from '../../../utils/dateUtil';
 import useSyncState from '../../../hooks/useSyncState';
 import type { BaseInfo, FormatType, Locale, ReplaceListType } from '../../../interface';
-import { fillIndex } from '../../util';
+import { formatValue, isSame, isSameTimestamp } from '../../../utils/dateUtil';
+import { fillIndex } from '../../../utils/miscUtil';
 import type { RangePickerProps } from '../RangePicker';
 import type { ReplacedPickerProps } from '../SinglePicker';
 import useLockEffect from './useLockEffect';
