+++ src/utils/miscUtil.ts
@@ -16,3 +16,10 @@ export function toArray<T>(val: T | T[]): T[] {
 
   return Array.isArray(val) ? val : [val];
 }
+
+export function fillIndex<T extends any[]>(ori: T, index: number, value: T[number]): T {
+  const clone = [...ori] as T;
+  clone[index] = value;
+
+  return clone;
+}
